Lots of work on IpNetwork
To summarize: The IpNetwork is now its own object that does not need any API in order to work. This is trying to mirror the same things that the WirelessNetwork is doing. This allows for the Network class to remain a singleton that can do everything it needs to without the IpNetwork having its own strange extra network that might come about through extending. The item IpNetworkCard has been added and has a dummy function that returns a 0.0.0.0 address that is called getIpNetwork. There are also some parts for things like achivements, tool tips, recipies, etc.

HOW TO ADD NEW FUNCTIONS TO IPNETWORKCARD
to add a new function follow this format:
@Callback(doc="""explanation""")
def name(context: Context, args: Arguments): Array[AnyRef] = {
    result()
  }

@Callback adds the function to the list of usable functions in game
doc provides the tip in game
name is the name of the function
context and args are how you get the different arguments that you might want from the function
the return values go into the result argument list so that they can be given back to the game